requirements.txt
streamlit==1.38.0
tensorflow==2.12.1
numpy==1.23.5
pandas==2.0.3
pillow==10.4.0


